requirements.txt
flask
pandas
pillow
requests
icrawler
